// src/enum.ts
export enum Key {
  Up = 'ArrowUp',
  Right = 'ArrowRight',
  Down = 'ArrowDown',
  Left = 'ArrowLeft',
  W = 'KeyW',
  D = 'KeyD',
  S = 'KeyS',
  A = 'KeyA',
  Enter = 'Enter',
}

export enum Direction {
  Up = 'Up',
  Right = 'Right',
  Down = 'Down',
  Left = 'Left',
}

export enum EnergySize {
  Small = 7,
  Medium = 12,
  Large = 18,
}

export enum Skill {
  Shield = 'shield',
  Speed = 'speed',
  Reload = 'reload',
  Bullet = 'bullet',
}